Handle python3 string with hash
