+++ ckanext/datajson/datajson_ckan_28.py
@@ -782,8 +782,8 @@ class DatasetHarvesterBase(HarvesterBase):
                                 "|" + harvest_source.config + "|" +  # NOQA W504
                                 self.HARVESTER_VERSION).hexdigest()
         else:
-            return hashlib.sha1(json.dumps(datasetdict, sort_keys=True) + "|" + json.dumps(catalog_extras,
-                                sort_keys=True)).hexdigest()
+            return hashlib.sha1((json.dumps(datasetdict, sort_keys=True) + "|" + json.dumps(catalog_extras,
+                                sort_keys=True)).encode('utf-8')).hexdigest()
 
     def find_extra(self, pkg, key):
         for extra in pkg["extras"]:
